requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.0
httpx==0.27.0
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.0
torch==2.3.0
transformers==4.41.2
pillow==10.4.0
python-dotenv==1.0.1
